Removed contruction banner and cleaned up layouts
