+++ src/components/hobbies/SpotifyCard.js
@@ -6,8 +6,6 @@ const SpotifyCard = (props) => {
 		<StyledWrapper>
 			<iframe
 				src={`https://open.spotify.com/embed/album/${props.trackId}`}
-				width='300'
-				height='380'
 				frameborder='0'
 				allowtransparency='true'
 				allow='encrypted-media'></iframe>
@@ -19,15 +17,14 @@ export default SpotifyCard;
 
 const StyledWrapper = styled.div`
 	width: 100%;
-	height: 300px;
+	height: 200px;
 	border-radius: 10px;
 	padding: 0;
 	margin: 0 16px 16px 0;
 
 	iframe {
 		width: 100%;
-		height: 100%;
+		height: 300px;
 		border-radius: 10px;
-		box-shadow: var(--nav-shadow);
 	}
 `;
