+++ src/components/layout.js
@@ -4,20 +4,14 @@ import Footer from './footer/footer';
 import PropTypes from 'prop-types';
 import React from 'react';
 import Sidebar from './sidebar/sidebar';
-import { UnderConstructionBanner } from './UnderConstructionBanner';
 import { ThemeProvider } from 'styled-components';
 import { theme } from '../contexts/StyleProvider';
-import styled from 'styled-components';
 
-const Layout = ({ children, isUnderConstruction }) => {
+const Layout = ({ children }) => {
 	return (
 		<ThemeProvider theme={theme}>
 			<Sidebar />
-			<main>
-				{isUnderConstruction && <UnderConstructionBanner />}
-				{isUnderConstruction && <StyledSpacer />}
-				{children}
-			</main>
+			<main>{children}</main>
 			<Footer />
 		</ThemeProvider>
 	);
@@ -28,9 +22,3 @@ Layout.propTypes = {
 };
 
 export default Layout;
-
-const StyledSpacer = styled.div`
-	@media only screen and (max-width: 600px) {
-		padding: 16px 0;
-	}
-`;
